+++ ninja-servlet-integration-test/src/main/java/conf/Routes.java
@@ -141,7 +141,7 @@ public class Routes implements ApplicationRoutes {
             router.GET().route("/_test/testPage").with(ApplicationController.class, "testPage");
         }
 
-        router.GET().route("/assets/webjars/{fileName: .*}").with(AssetsController.class, "serveWebjars");
+        router.GET().route("/assets/webjars/{fileName: .*}").with(AssetsController.class, "serveWebJars");
         router.GET().route("/assets/{fileName: .*}").with(AssetsController.class, "serveStatic");
         
     }